# internal/spec/swagger.yaml
basePath: /api/v1
definitions:
  github_com_new-marty_health-connect_internal_apperror.Body:
    properties:
      code:
        $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.ErrorCode'
      details:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  github_com_new-marty_health-connect_internal_apperror.ErrorCode:
    enum:
    - NOT_FOUND
    - INVALID_INPUT
    - INTERNAL_ERROR
    - TIMEOUT
    type: string
    x-enum-varnames:
    - CodeNotFound
    - CodeInvalidInput
    - CodeInternal
    - CodeTimeout
  github_com_new-marty_health-connect_internal_apperror.Response:
    properties:
      error:
        $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Body'
    type: object
...
  internal_sync.SyncStatus:
    properties:
      last_status:
        type: string
      last_sync_at:
        type: string
      records_added:
        type: integer
      source:
        type: string
    type: object
info:
  contact: {}
  description: Personal health data API. Aggregates Oura, Strava, Hevy, InBody, Apple
    Health, and meal logs. Every endpoint is also exposed as an LLM tool — see /openapi.json,
    /tools/openai.json, /tools/anthropic.json.
  title: Health Connect API
  version: "1.0"
paths:
...
  /sync/{source}:
    post:
      description: Forces a sync for the named source. Pass "all" to sync every registered
        source. Use when the user wants their data refreshed before analysis.
      operationId: sync_trigger
      parameters:
      - description: 'Source name: oura, strava, hevy, inbody, or all'
        in: path
        name: source
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              records_added:
                type: integer
              source:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
      security:
      - BearerAuth: []
      summary: Trigger an on-demand sync
      tags:
      - sync
  /sync/status:
    get:
      description: Returns last successful sync timestamp, last status, and total
        records added per source.
      operationId: sync_status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/internal_sync.SyncStatus'
                type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_new-marty_health-connect_internal_apperror.Response'
      security:
      - BearerAuth: []
      summary: Get sync status per source
      tags:
      - sync
securityDefinitions:
  BearerAuth:
    description: Bearer token. Set the same value in HEALTH_CONNECT_API_TOKEN on the
      server.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"